Allow the user to play with the rules without killing the program.
This was fun. Uploading for later. I'm not sure the right way to search
the space, but I was loosely trying to figure out how to get the grid to
"send" something across the way, like a walker, but also allow chaos to
create pockets of interest without it being pre-designed that way
(randomness will create patterns).